+++ Cargo.toml
@@ -24,7 +24,7 @@ is-it-maintained-open-issues = { repository = "enarx/enarx" }
 [features]
 default = ["backend-kvm", "backend-sgx", "backend-sev", "wasmldr"]
 
-backend-sev = ["backend-kvm", "reqwest"]
+backend-sev = ["backend-kvm", "ureq"]
 backend-kvm = ["x86_64", "kvm-bindings", "kvm-ioctls"]
 backend-sgx = ["x86_64", "sgx"]
 wasmldr = []
@@ -59,7 +59,7 @@ vdso = "0.2"
 log = "0.4"
 spinning = "0.1.0"
 env_logger = "0.9"
-reqwest = { version = "0.11", features = [ "blocking" ], optional = true }
+ureq = { version = "2.4.0", optional = true }
 serde = { version = "1.0.136", features = ["derive"] }
 serde_json = "1.0.79"
 atty = "0.2"
