+++ deny.toml
@@ -10,11 +10,10 @@ unlicensed = "deny"
 allow = [
     "Apache-2.0",
     "Apache-2.0 WITH LLVM-exception",
-    "BSD-2-Clause",
     "BSD-3-Clause",
-    "CC0-1.0",
     "ISC",
     "MIT",
+    "MPL-2.0",
     "OpenSSL",
     "Zlib",
 ]
